Generate libksi.spec with VERSION number in it
If one wants to build the libksi RPMs having just the source tar ball
and the libksi.spec file (which is what you get when trying to build
libksi using the source RPM), the .spec file does not have the VERSION
file around yet to get the necessary value.

By building the .spec file explicitly as part of the "make dist" step,
then we'll have a libksi.spec file which when paired with the generated
tar ball can be built independently of the original source tree.